requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF 处理
pdfplumber==0.10.3
pypdfium2==4.25.0
pdf2image==1.16.3
pypdf==3.17.4

# LLM 客户端
openai==1.3.7
dashscope==1.17.0

# 图片处理
Pillow==10.1.0
html2image==2.2.0
playwright==1.40.0

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# 其他
jinja2==3.1.2
markdown==3.5.1

